Add support for "Hollow Palm Technique" Keystone (#1206)
* Implement generic "per x `defence` on all armour"

* Remove PoE1 "Hollow Palm" from `ModParser` code

* Implement "as though using ..." weapon type check

"Hollow Palm" introduces the "as though using a Quarterstaff" wording.
Current interpretation is that this is different from "count as using"
wording and only removes weapon type restrictions. This commit adds
support for that kind of behavior.

Note:
Because "WeaponData" mods are only processed in `Item.lua` and only for
actual items. This kind of mod will not work via `ModParser` when
modifying "Unarmed" weaponData

* Change "Hollow Palm" processing in `CalcSetup`

TODO:
This uses "Eternal Youth" as a placeholder Keystone for testing because
"Hollow Palm" is not yet available in 0.2.

* Disable Heralds and Persistent Buffs

The assumption is that Hollow Palm will allow Quartestaff Attack skills,
but not make you count as wielding a "Martial Weapon" for things like
Heralds and Wind Dancer.

TODO:
- Update to use actual mod wording once released
- Check if assumption is true in game

* Remove lines disabling Heralds and Wind Dancer

* Add preliminary "added phys based on gem level"

* Update flat phys values with game data

* Implement actual mod parsing for full wording

* Update placeholder Keystone name

* Use data exported from ggpk instead

Flat phys data for Hollow Palm Technique is stored in
`flatphysicaldamagevalues.dat64` which is now being exported to
`Misc.lua` and used instead of the hardcoded table in `Data.lua`

* Remove unicode bullet points

---------

Co-authored-by: majochem <majochem@users.noreply.github.com>
Co-authored-by: LocalIdentity <localidentity2@gmail.com>