Add support for Totems (#1188)
* Add 'totem' flag to Vile Effusion (Dark Effigy)

The 'totem' flag was missing for the damaging part of the skill

* Fix `addFlags` parsing

`addFlags` functionality was accidentally broken in #62 because
`mainSkillFlags` and `calcSkillFlags` variables were introduced but
seemingly later made redundant to use the original `skillFlags`.

However the `addFlags` section still checked the old variables which were
now `nil`. This wasn't a problem until now, because the PoE2 skills had
not used `addFlags` yet.

* Remove "Base" mod for "ActiveTotemLimit"

`CalcSetup` previous added +1 to "ActiveTotemLimit" as a base, but
totem skills in PoE2 always come with "base_number_of_totems_allowed"
values, which led to the values always being 1 too high

* Let "Ancestral Warrior Totem" add the `totem` flag

* Get totem stats from "base" totem skill

This fixes previous problems that were caused by totem skills consisting
of multiple parts, such as inability to get correct skillTotemId and
level requirement for the "active" portion of a totem skill.

The skill (or support) that provides the baseTotem data is now
identified within `CalcActiveSkill`

Examples:
- "Shockwave Totem" -> "Wave" would not show correct totem stats
- "Sunder" attached to "Ancestral Warrior Totem" would not show any
  totem stats at all

* Reword "totemification" to "totemified"

* Get rid of "totemified"

* Add support for Mortar Cannon

Adds support for a few stats on mortar cannon and ancestral totem too

---------

Co-authored-by: majochem <majochem@users.noreply.github.com>
Co-authored-by: LocalIdentity <localidentity2@gmail.com>