Fix Abyss Timeless jewel mod display (#1365)
The mods granted by the timeless jewel do not appear in the node tooltip cause it is pulled from the tree file instead of the modified one. Not sure if this will cause bugs elsewhere
The change in CalcSetup is needed so that it doesn't run the whole calc setup file every time you allocate a node but it causes the issue where stat don't get added to the base attribute nodes. Not sure how to fix this

Co-authored-by: LocalIdentity <localidentity2@gmail.com>